提交占比除以100;

+++ src/com/dkhs/portfolio/ui/adapter/OptionalStockAdapter.java
@@ -99,7 +99,7 @@ public class OptionalStockAdapter extends BaseAdapter {
 
         GradientDrawable gd = (GradientDrawable) sd.getDrawable();
         gd.setColor(item.getDutyColor());
-        tvPercent.setText(StringFromatUtils.getPercentValue((int) (item.getPercent())));
+        tvPercent.setText(StringFromatUtils.getPercentValue((item.getPercent())));
 
         seekbar.setOnSeekBarChangeListener(new OnSeekBarChangeListener() {
             @Override
